user Verification from the log in form moved to a different verifyUser helper
